requirements.txt
streamlit
pandas
pillow
torch
timm
transformers
numpy
requests
Image
